+++ app/src/main/java/com/flowfoundation/wallet/page/wallet/model/WalletHeaderModel.kt
@@ -2,12 +2,13 @@ package com.flowfoundation.wallet.page.wallet.model
 
 import com.flowfoundation.wallet.network.model.WalletListData
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class WalletHeaderModel(
     @SerializedName("walletList")
     val walletList: WalletListData,
     @SerializedName("balance")
-    var balance: Float,
+    var balance: BigDecimal,
     @SerializedName("coinCount")
     var coinCount: Int = 0,
     @SerializedName("transactionCount")
